+++ busy_beaver/blueprints/__init__.py
@@ -1,4 +1,4 @@
 from .healthcheck import healthcheck_bp  # noqa
-from .poller import poller_bp  # noqa
 from busy_beaver.apps.github_integration.api import github_bp  # noqa
+from busy_beaver.apps.poller.api import poller_bp  # noqa
 from busy_beaver.apps.slack_integration.api import slack_bp  # noqa
